// restaurant-app/src/app/model/chart/chart-name.ts
export enum ChartName {
  COMPARE_ORDERS_AMOUNT = 'Liczba zamówień online w stosunku do liczby zamówień na miejscu',
  COMPARE_ORDERS_INCOME = 'Przychód z zamówień online w stosunku do liczby zamówień na miejscu',
  ORDERS_AMOUNT_USING_DIFFERENT_PAYMENT_METHODS = 'Liczba zamówień z użyciem różnych metod płatności',
  ORDERS_AMOUNT_IN_DIFFERENT_HOURS = 'Liczba zamówień w zależności od godziny',
  SOLD_DISH_AMOUNT = 'Liczba sprzedanych poszczególnych potraw',
  SOLD_DISH_INCOME = 'Zrobek ze sprzedanych poszczególnych potraw',
  AVERAGE_COMPLETION_ORDER_TIME_DEPENDS_ON_DISHES_AMOUNT = 'Średni czas realizacji zamówienia w zależności od ilości dań',
  ORDERS_AMOUNT_CONSIDERING_DISHES_AMOUNT_IN_EACH = 'Ilość zamówień pod względem potraw w zamówienia',
}